backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
motor==3.6.0
bcrypt==4.2.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
opencv-python-headless==4.10.0.84
numpy==2.1.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
